backend/requirements.txt
google-genai>=0.8.0
google-adk>=1.26.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
websockets>=12.0
google-cloud-firestore>=2.16.0
google-cloud-storage>=2.14.0
python-dotenv>=1.0.0
pydantic>=2.5.0
